+++ graphics/r_window.h
@@ -13,5 +13,5 @@ namespace Fresa::Graphics::Window
     WindowData create(Vec2<> size, str name);
     ui16 getRefreshRate(WindowData &win, bool force = false);
     
-    inline Events::Event<Vec2<>> event_window_resize;
+    inline Event::Event<Vec2<>> event_window_resize;
 }